[COS-21][Webscrpaer] - replace ai module usage with common ai module
