+++ packages/server/customer-os-common-ai/go.mod
@@ -5,9 +5,9 @@ go 1.20
 require (
 	github.com/opentracing/opentracing-go v1.2.0
 	github.com/sirupsen/logrus v1.9.3
+	github.com/google/go-querystring v1.1.0
 )
 
 require (
-	github.com/google/go-querystring v1.1.0
 	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
 )
