--- packages/server/customer-os-common-ai/service/anthropic/config.go
@@ -1,6 +1,0 @@
-package anthropic
-
-type AiModelConfigAnthropic struct {
-	ApiKey  string `json:"apiKey"`
-	ApiPath string `json:"apiPath"`
-}
