--- packages/server/customer-os-common-ai/service/openai/config.go
@@ -1,7 +1,0 @@
-package openai
-
-type AiModelConfigOpenAi struct {
-	ApiKey       string `json:"apiKey"`
-	Organization string `json:"organization"`
-	Model        string `json:"model"`
-}
